requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
boto3==1.34.0
Pillow==10.1.0
qrcode[pil]==7.4.2